Add files via upload
Data Science extracts meaningful insight from chunks of raw data, which is
useful to different business segments for planning their future course of action. Finance is probably

one of the first to catch on to this trend with a rise in the penetration of analytics into many aspects of
our lives. Here, we will analyze data from the stock market for some technology stocks such as
Apple, Google, Amazon, and Microsoft
.Use Python libraries such as Pandas, Seaborn, and Matplotlib to extract and analyze the
information, visualize it, and look at different ways to analyze the risk of a stock, based on its
performance history.